@@ -1,12 +1,12 @@
 name: facebook_all_in_one
 description: A new flutter plugin project.
-version: 0.0.15+22
+version: 0.1.0+27
 author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.12.0 <2.0.0"
 
 dependencies:
   flutter:
